+++ Medilink-Final-Project/Areas/Admin/Controllers/Home/HomeSpecialityModernController.cs
@@ -66,6 +66,11 @@ namespace Medilink_Final_Project.Areas.Admin.Controllers.Home
                     ModelState.AddModelError("Upload", "Siz yalnız png,jpg və ya gif faylı yükləyə bilərsiniz");
                 }
 
+                if (homeModern.Upload.Length > 1048576)
+                {
+                    ModelState.AddModelError("Upload", "Fayl ölcüsu maximum 1MB ola bilər");
+                }
+
             }
             if (ModelState.IsValid)
             {
@@ -116,6 +121,13 @@ namespace Medilink_Final_Project.Areas.Admin.Controllers.Home
                             ModelState.AddModelError("Upload", "Siz yalnız png,jpg və ya gif faylı yükləyə bilərsiniz");
                             return View(homeModern);
                         }
+
+                        if (homeModern.Upload.Length > 1048576)
+                        {
+                            ModelState.AddModelError("Upload", "Fayl ölcüsu maximum 1MB ola bilər");
+                            return View(homeModern);
+                        }
+
                         var oldFile = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "uploads", homeModern.Photo);
                         _fileManager.Delete(oldFile);
